add more bgm add cpp
