+++ Key.h
@@ -1,0 +1,6 @@
+﻿#pragma once
+#include<Windows.h>
+
+bool key_down(const unsigned char& key_value);
+
+bool key_short_press(const unsigned char& key_value);
