+++ main.cpp
@@ -6,14 +6,14 @@ int main()
 	std::cin.tie(NULL);
 	std::cout.tie(NULL);
 
-	srand(time(NULL));
+	srand((unsigned int)time(NULL));
 
 	cursor_hide();
 
 	while (true)
 	{
-		mciSendStringW(L"open ./music/1.mp3 alias bgm", NULL, 0, NULL);
-		mciSendStringW(L"play bgm", NULL, 0, NULL);
+		mciSendStringW(L"open ./music/1.mp3 alias bgm1", NULL, 0, NULL);
+		mciSendStringW(L"play bgm1", NULL, 0, NULL);
 		system("cls");
 		print_menu();
 		int opt;
@@ -24,11 +24,14 @@ int main()
 			print_description();
 			int difficuty;
 			std::cin >> difficuty;
+			mciSendStringW(L"close bgm2", NULL, 0, NULL);
 			print_difficuty(difficuty);
+			mciSendStringW(L"open ./music/Kirov.mp3 alias Kirov", NULL, 0, NULL);
+			mciSendStringW(L"play Kirov", NULL, 0, NULL);
 			Sleep(2000);
 			if (difficuty == 1)
 			{
-				play_game(25, 115, 300, -1, -1);
+				play_game(25, 115, 100, -1, -1);
 			}
 			if (difficuty == 2)
 			{
@@ -45,6 +48,6 @@ int main()
 		}
 		if (opt == 2) break;
 	}
-	mciSendStringW(L"close bgm", NULL, 0, NULL);
+	mciSendStringW(L"close bgm1", NULL, 0, NULL);
 	return 0;
 }
