+++ main.h
@@ -3,6 +3,6 @@
 #include<Windows.h>
 #include "Time.h"
 #include "Cls_and_Print.h"
-#include "Play_Game.h"
+#include "Game.h"
 
 #pragma comment(lib, "winmm.lib")
